补充多用户场景无障碍切换的tdd用例.Co-Authored-by: Agent.Signed-off-by: <zhouke35@huawei.com>.

+++ testserver/src/service/test_server_service.cpp
@@ -689,7 +689,7 @@ namespace OHOS::testserver {
             counts = 1;
             return TEST_SERVER_ACCOUNTOP_FAILED;
         }
-        counts = activatedOsAccountIds.size();
+        counts = static_cast<int32_t>(activatedOsAccountIds.size());
         return TEST_SERVER_OK;
     }
